+++ R/techdatareport_spec.R
@@ -28,4 +28,4 @@
 #'
 #' @family Data Specs
 #'
-"maintrepair_spec"
+"techdatareport_spec"