+++ R/flexfile_family-class.R
@@ -1,6 +1,10 @@
 
 #' @keywords internal
-new_flexfile_family <- function(flexfile = NULL, quantityreport = NULL, maintrepair = NULL, techdatareport = NULL) {
+new_flexfile_family <- function(flexfile = NULL,
+                                quantityreport = NULL,
+                                maintrepair = NULL,
+                                techdatareport = NULL,
+                                .check = TRUE) {
 
   report_list <- list(flexfile = flexfile,
                       quantityreport = quantityreport,
@@ -9,8 +13,10 @@ new_flexfile_family <- function(flexfile = NULL, quantityreport = NULL, maintrep
 
   report_list <- purrr::compact(report_list)
 
-  check_reports <- identical_report_metadata(report_list = report_list)
-  if (!check_reports) stop("Each report must be from the same FlexFile family.")
+  if (isTRUE(.check)) {
+    check_reports <- identical_report_metadata(report_list = report_list)
+    if (!check_reports) stop("Each report must be from the same FlexFile family.")
+  }
 
   structure(report_list, class = "flexfile_family")
 }